+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Please see each crate's change log below:
 
+- [`anyspawn`](./crates/anyspawn/CHANGELOG.md)
 - [`bytesbuf`](./crates/bytesbuf/CHANGELOG.md)
 - [`bytesbuf_io`](./crates/bytesbuf_io/CHANGELOG.md)
 - [`data_privacy`](./crates/data_privacy/CHANGELOG.md)